rename: 색상 파일 이름 수정 및 색상 가져오기 경로 수정
